Extract `sql` module with all custom `sql_function!()` calls
